finetune/requirements.txt
transformers>=4.45.0
peft>=0.13.0
trl>=0.12.0
datasets>=3.0.0
accelerate>=0.34.0
huggingface_hub>=0.24.0
scipy>=1.11.0
sentencepiece>=0.2.0
protobuf>=4.25.0
pyyaml>=6.0
torch>=2.4.0